completed Complex class

+++ Fract.h
@@ -4,8 +4,9 @@
 #include <iostream>
 #include <math.h>
 using namespace std;
+
 class Fract {
-    int tu, mau;
+    double tu, mau;
 public:
     int getNumer();
     int getDeno();
@@ -20,5 +21,6 @@ public:
     Fract operator/(Fract&);
     friend istream& operator>>(istream&, Fract&);
     friend ostream& operator<<(ostream&, const Fract&);
+    
 };
 #endif
